requirements.txt
crewai>=0.55.0
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
